requirements.txt
Pillow
python-telegram-bot
python-dotenv